refactor: 💡 improve on locations and travels services/resolvers

+++ src/graphql/modules/locations/locations.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, UnauthorizedException } from '@nestjs/common';
 
 import { PrismaService } from '../../../database/prisma/prisma.service';
 import { CreateLocationInput, UpdateLocationInput } from '../../inputs/location.input';
@@ -17,7 +17,13 @@ export class LocationsService {
     });
   }
 
-  async createLocation(data: CreateLocationInput) {
+  async getMyLocations(userId: string) {
+    return await this.prisma.location.findMany({
+      where: { createdByUserId: userId },
+    });
+  }
+
+  async createLocation(data: CreateLocationInput, userId: string) {
     const isLocationAlreadyExists = await this.prisma.location.findFirst({
       where: { country: data.country, city: data.city },
     });
@@ -25,18 +31,39 @@ export class LocationsService {
     if (isLocationAlreadyExists) throw new BadRequestException('Location already exists');
 
     return await this.prisma.location.create({
-      data,
+      data: {
+        ...data,
+        createdByUserId: userId,
+      },
     });
   }
 
-  async updateLocation(data: UpdateLocationInput) {
+  async updateLocation(data: UpdateLocationInput, userId: string) {
+    if (userId) throw new BadRequestException('You must provide userId to updating a location');
+
+    const isUserAuthorizedToUpdateLocation = await this.prisma.location.findFirst({
+      where: { id: data.id, createdByUserId: userId },
+    });
+
+    if (!isUserAuthorizedToUpdateLocation)
+      throw new UnauthorizedException("You aren't authorized to update this location, or location doesn't exist");
+
     return await this.prisma.location.update({
       where: { id: data.id },
       data,
     });
   }
 
-  async deleteLocation(id: string) {
+  async deleteLocation(id: string, userId: string) {
+    if (!userId) throw new BadRequestException('You must have be authenticated to can delete the location');
+
+    const isUserAuthorizedToDeleteLocation = await this.prisma.location.findFirst({
+      where: { id, createdByUserId: userId },
+    });
+
+    if (!isUserAuthorizedToDeleteLocation)
+      throw new UnauthorizedException("You aren't authorized to delete this location, or location doesn't exist");
+
     return await this.prisma.location.delete({
       where: { id },
     });
